🎨 Refactoring + started working on logger
