+++ Gemfile
@@ -2,6 +2,7 @@ source "https://rubygems.org"
 gemspec
 gem 'sorbet', :group => :development
 gem 'sorbet-runtime'
+gem 'concurrent-ruby', '~> 1.3', '>= 1.3.3'
 gem 'tapioca', require: false, :group => [:development, :test]
 gem 'simplecov', require: false, :group => [:development, :test]
 gem 'simplecov-cobertura', require: false, :group => [:development, :test]
