--- lib/arbac_verifier/exceptions/computation_timed_out_exception.rb
@@ -1,4 +1,0 @@
-# typed: strict
-# Public: Specific exception to throw when a certain computation time exceeds a predefined limit
-class ComputationTimedOutException < StandardError
-end
