+++ sorbet/rbi/hidden-definitions/hidden.rbi
@@ -3,18 +3,33 @@
 
 # typed: autogenerated
 
+class ARBACVerifier::Instance
+  extend ::T::Private::Methods::MethodHooks
+  extend ::T::Private::Methods::SingletonMethodHooks
+end
+
+class ARBACVerifier::Rules::CanAssign
+  extend ::T::Private::Methods::MethodHooks
+  extend ::T::Private::Methods::SingletonMethodHooks
+end
+
+class ARBACVerifier::Rules::CanRevoke
+  extend ::T::Private::Methods::MethodHooks
+  extend ::T::Private::Methods::SingletonMethodHooks
+end
+
+class ARBACVerifier::UserRole
+  extend ::T::Private::Methods::MethodHooks
+  extend ::T::Private::Methods::SingletonMethodHooks
+end
+
+module ARBACVerifier::Utils
+  extend ::T::Private::Methods::MethodHooks
+  extend ::T::Private::Methods::SingletonMethodHooks
+end
+
 class Addrinfo
   def connect_internal(local_addrinfo, timeout=T.unsafe(nil)); end
-end
-
-class ArbacInstance
-  extend ::T::Private::Methods::MethodHooks
-  extend ::T::Private::Methods::SingletonMethodHooks
-end
-
-module ArbacUtilsModule
-  extend ::T::Private::Methods::MethodHooks
-  extend ::T::Private::Methods::SingletonMethodHooks
 end
 
 class Array
@@ -2629,18 +2644,110 @@ end
 module CGI::HtmlExtension
 end
 
-class CanAssignRule
-  extend ::T::Private::Methods::MethodHooks
-  extend ::T::Private::Methods::SingletonMethodHooks
-end
-
-class CanRevokeRule
-  extend ::T::Private::Methods::MethodHooks
-  extend ::T::Private::Methods::SingletonMethodHooks
-end
-
 class Class
   def json_creatable?(); end
+end
+
+module Concurrent
+  NULL = ::T.let(nil, ::T.untyped)
+  NULL_LOGGER = ::T.let(nil, ::T.untyped)
+  VERSION = ::T.let(nil, ::T.untyped)
+end
+
+class Concurrent::AbstractExecutorService
+  FALLBACK_POLICIES = ::T.let(nil, ::T.untyped)
+end
+
+Concurrent::Collection::MapImplementation = Concurrent::Collection::MriMapBackend
+
+class Concurrent::ConcurrentUpdateError
+  CONC_UP_ERR_BACKTRACE = ::T.let(nil, ::T.untyped)
+end
+
+class Concurrent::FiberLocalVar
+  LOCALS = ::T.let(nil, ::T.untyped)
+end
+
+class Concurrent::LockFreeStack
+  EMPTY = ::T.let(nil, ::T.untyped)
+end
+
+Concurrent::LockLocalVar = Concurrent::FiberLocalVar
+
+class Concurrent::MVar
+  EMPTY = ::T.let(nil, ::T.untyped)
+  TIMEOUT = ::T.let(nil, ::T.untyped)
+end
+
+class Concurrent::Maybe
+  NONE = ::T.let(nil, ::T.untyped)
+end
+
+module Concurrent::Promises::InternalStates
+  PENDING = ::T.let(nil, ::T.untyped)
+  RESERVED = ::T.let(nil, ::T.untyped)
+  RESOLVED = ::T.let(nil, ::T.untyped)
+end
+
+class Concurrent::ReadWriteLock
+  MAX_READERS = ::T.let(nil, ::T.untyped)
+  MAX_WRITERS = ::T.let(nil, ::T.untyped)
+  RUNNING_WRITER = ::T.let(nil, ::T.untyped)
+  WAITING_WRITER = ::T.let(nil, ::T.untyped)
+end
+
+class Concurrent::ReentrantReadWriteLock
+  MAX_READERS = ::T.let(nil, ::T.untyped)
+  MAX_WRITERS = ::T.let(nil, ::T.untyped)
+  READER_BITS = ::T.let(nil, ::T.untyped)
+  READ_LOCK_MASK = ::T.let(nil, ::T.untyped)
+  RUNNING_WRITER = ::T.let(nil, ::T.untyped)
+  WAITING_WRITER = ::T.let(nil, ::T.untyped)
+  WRITER_BITS = ::T.let(nil, ::T.untyped)
+  WRITE_LOCK_HELD = ::T.let(nil, ::T.untyped)
+  WRITE_LOCK_MASK = ::T.let(nil, ::T.untyped)
+end
+
+class Concurrent::RubyThreadPoolExecutor
+  DEFAULT_MAX_POOL_SIZE = ::T.let(nil, ::T.untyped)
+  DEFAULT_MAX_QUEUE_SIZE = ::T.let(nil, ::T.untyped)
+  DEFAULT_MIN_POOL_SIZE = ::T.let(nil, ::T.untyped)
+  DEFAULT_SYNCHRONOUS = ::T.let(nil, ::T.untyped)
+  DEFAULT_THREAD_IDLETIMEOUT = ::T.let(nil, ::T.untyped)
+end
+
+class Concurrent::SerializedExecutionDelegator
+  RUBYGEMS_ACTIVATION_MONITOR = ::T.let(nil, ::T.untyped)
+end
+
+class Concurrent::SynchronizedDelegator
+  RUBYGEMS_ACTIVATION_MONITOR = ::T.let(nil, ::T.untyped)
+end
+
+class Concurrent::ThreadLocalVar
+  LOCALS = ::T.let(nil, ::T.untyped)
+end
+
+module Concurrent::ThreadSafe::Util
+  CPU_COUNT = ::T.let(nil, ::T.untyped)
+  FIXNUM_BIT_SIZE = ::T.let(nil, ::T.untyped)
+  MAX_INT = ::T.let(nil, ::T.untyped)
+end
+
+class Concurrent::TimerTask
+  DEFAULT_INTERVAL_TYPE = ::T.let(nil, ::T.untyped)
+  EXECUTION_INTERVAL = ::T.let(nil, ::T.untyped)
+  FIXED_DELAY = ::T.let(nil, ::T.untyped)
+  FIXED_RATE = ::T.let(nil, ::T.untyped)
+end
+
+class Concurrent::Transaction
+  ABORTED = ::T.let(nil, ::T.untyped)
+end
+
+module Concurrent::Utility::NativeInteger
+  MAX_VALUE = ::T.let(nil, ::T.untyped)
+  MIN_VALUE = ::T.let(nil, ::T.untyped)
 end
 
 module Coverage
@@ -4219,6 +4326,32 @@ class LoadError
   include ::DidYouMean::Correctable
 end
 
+class Logger
+  def debug!(); end
+
+  def error!(); end
+
+  def fatal!(); end
+
+  def info!(); end
+
+  def warn!(); end
+  SEV_LABEL = ::T.let(nil, ::T.untyped)
+end
+
+class Logger::Formatter
+  DatetimeFormat = ::T.let(nil, ::T.untyped)
+  Format = ::T.let(nil, ::T.untyped)
+end
+
+class Logger::LogDevice
+  include ::MonitorMixin
+end
+
+module Logger::Period
+  SiD = ::T.let(nil, ::T.untyped)
+end
+
 class MatchData
   def match(arg); end
 
@@ -4509,11 +4642,6 @@ end
 RSpec::Core::Example::AllExceptionsExcludingDangerousOnesOnRubiesThatAllowIt = RSpec::Support::AllExceptionsExceptOnesWeMustNotRescue
 
 class RSpec::Core::ExampleGroup
-  include ::RSpec::Core::MockingAdapters::RSpec
-  include ::RSpec::Mocks::ExampleMethods
-  include ::RSpec::Mocks::ArgumentMatchers
-  include ::RSpec::Mocks::ExampleMethods::ExpectHost
-  include ::RSpec::Matchers
   INSTANCE_VARIABLE_TO_IGNORE = ::T.let(nil, ::T.untyped)
 end
 
@@ -5197,6 +5325,10 @@ class Thread
   def self.new(*arg); end
 end
 
+module Timeout
+  VERSION = ::T.let(nil, ::T.untyped)
+end
+
 class TracePoint
   def eval_script(); end
 
@@ -5314,11 +5446,6 @@ module UnicodeNormalize
 end
 
 module UnicodeNormalize
-end
-
-class UserRole
-  extend ::T::Private::Methods::MethodHooks
-  extend ::T::Private::Methods::SingletonMethodHooks
 end
 
 module Warning
