+++ spec/unit/classes/instance_spec.rb
@@ -1,14 +1,14 @@
 # typed: false
 # frozen_string_literal: true
-require 'arbac_verifier/classes/arbac_instance'
+require 'arbac_verifier/classes/instance'
 require 'rspec'
 
-describe ArbacInstance do
+describe ARBACVerifier::Instance do
 
   describe '#initialize' do
     context 'given a file path' do
       it 'parses successfully the given file and returns a valid arbac instance' do
-        arbac_instance = ArbacInstance.new(path: "spec/fixtures/policies/policy0.arbac")
+        arbac_instance = ARBACVerifier::Instance.new(path: "spec/fixtures/policies/policy0.arbac")
 
         expect(arbac_instance).to_not be_nil
         expect(arbac_instance.goal).to be(:Student)
@@ -32,17 +32,17 @@ describe ArbacInstance do
 
     context 'given attributes' do
       it 'sets the given attributes in the new arbac instance' do
-        arbac_instance = ArbacInstance.new(
+        arbac_instance = ARBACVerifier::Instance.new(
           goal: :Student,
           roles: Set.new([:Teacher, :Student, :TA]),
           users: Set.new(["stefano", "alice", "bob"]),
-          user_to_role: Set.new([UserRole.new("stefano", :Teacher), UserRole.new("alice", :TA)]),
+          user_to_role: Set.new([ARBACVerifier::UserRole.new("stefano", :Teacher), ARBACVerifier::UserRole.new("alice", :TA)]),
           can_assign_rules: Set.new([
-                                      CanAssignRule.new(:Teacher, [].to_set, [:Teacher, :TA].to_set, :Student),
-                                      CanAssignRule.new(:Teacher, [].to_set, [:Student].to_set, :TA),
-                                      CanAssignRule.new(:Teacher, [:TA].to_set, [:Student].to_set, :Teacher)
+                                      ARBACVerifier::Rules::CanAssign.new(:Teacher, [].to_set, [:Teacher, :TA].to_set, :Student),
+                                      ARBACVerifier::Rules::CanAssign.new(:Teacher, [].to_set, [:Student].to_set, :TA),
+                                      ARBACVerifier::Rules::CanAssign.new(:Teacher, [:TA].to_set, [:Student].to_set, :Teacher)
                                     ]),
-          can_revoke_rules: Set.new([CanRevokeRule.new(:Teacher, :Student), CanRevokeRule.new(:Teacher, :TA)])
+          can_revoke_rules: Set.new([ARBACVerifier::Rules::CanRevoke.new(:Teacher, :Student), ARBACVerifier::Rules::CanRevoke.new(:Teacher, :TA)])
         )
 
         expect(arbac_instance).to_not be_nil
@@ -67,7 +67,7 @@ describe ArbacInstance do
 
     context 'given no params' do
       it 'fails with an argument error' do
-        expect { ArbacInstance.new() }.to raise_error(TypeError)
+        expect { ARBACVerifier::Instance.new() }.to raise_error(TypeError)
       end
     end
   end
