+++ spec/unit/classes/user_role_spec.rb
@@ -3,8 +3,8 @@
 require 'arbac_verifier/classes/user_role'
 require 'rspec'
 
-describe UserRole do
-  let (:sut) { UserRole.new("alberto", :Boss) }
+describe ARBACVerifier::UserRole do
+  let (:sut) { ARBACVerifier::UserRole.new("alberto", :Boss) }
 
   describe '#initialize' do
     it 'should create a proper instance of UserRole' do
@@ -18,15 +18,15 @@ describe UserRole do
     context 'when two UserRole instances have same attributes' do
 
       it 'should return true' do
-        expect(sut == UserRole.new("alberto", :Boss)).to be(true)
+        expect(sut == ARBACVerifier::UserRole.new("alberto", :Boss)).to be(true)
       end
 
     end
     context 'when two UserRole instances have different attributes' do
 
       it 'should return false' do
-        expect(sut == UserRole.new("alberto", :Vice)).to be(false)
-        sut == UserRole.new("stefano", :Boss)
+        expect(sut == ARBACVerifier::UserRole.new("alberto", :Vice)).to be(false)
+        sut == ARBACVerifier::UserRole.new("stefano", :Boss)
       end
 
     end
